Use unsafe.Slice for slice-type conversions
Use `unsafe.Slice` for Go 1.17+, which is the new canonical method to convert a data slice between types without changing the underlying bytes. Add a file that is built for Go -1.16 if needed, too. Fixes #267.
